requirements.txt
Django (2.0.5)
django-widget-tweaks (1.4.3)
djangorestframework (3.8.2)
gunicorn (19.9.0)
pip (1.5.6)
psycopg2 (2.7.4)
pytz (2019.1)
setuptools (5.5.1)
whitenoise (4.1.2)
